LayoutTests/fast/images/animated-gif-body-outside-viewport-expected.txt

Make sure the background image of a body element that is outside the viewport is not animated

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS isBackgroundAnimated is false
Scroll down so that the body becomes visible.
PASS isBackgroundAnimated is true
PASS successfullyParsed is true

TEST COMPLETE
